MCP-1920: moved to common domain
