+++ api/src/main/java/gov/va/vro/api/requests/MasClaimDetailsRequest.java
@@ -3,8 +3,8 @@ package gov.va.vro.api.requests;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import gov.va.vro.api.model.mas.ClaimDetail;
-import gov.va.vro.api.model.mas.VeteranIdentifiers;
+import gov.va.vro.model.mas.ClaimDetail;
+import gov.va.vro.model.mas.VeteranIdentifiers;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Getter;
