+++ controller/src/main/java/gov/va/vro/controller/MasController.java
@@ -13,9 +13,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class MasController implements MasResource {
 
-  /**
-   * Initiate MAS integration
-   */
+  /** Initiate MAS integration */
   @Override
   public ResponseEntity<MasClaimResponse> notifyAutomatedClaimDetails(
       MasClaimDetailsRequest request) {
